backend/requirements.txt
Django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
imagekitio==2.2.8
psycopg2-binary==2.9.3
python-dotenv==0.19.2
django-anymail[sendinblue]==8.5
djangorestframework-camel-case
django-update-from-dict
django-cors-headers
django-jazzmin
django[argon2]
django-json-widget
colorlog